[Refactor] Domain 별로 foldering (#17) - auth   - domain     - Token - common   - baseEntities     - CreationLog     - CreationModificationLog - members   - domain     - Member - messages   - domain     - Message     - Report - missions   - domain     - CommonMission     - IndividualMission     - Mission     - MissionType - rooms   - domain     - MemberRoom     - ParticipantRole     - Relation     - Room     - RoomState
