+++ aenitto/src/main/java/com/firefighter/aenitto/common/baseEntities/CreationLog.java
@@ -1,8 +1,7 @@
-package com.firefighter.aenitto.domain;
+package com.firefighter.aenitto.common.baseEntities;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.Column;
@@ -13,12 +12,8 @@ import java.time.LocalDateTime;
 @MappedSuperclass
 @Getter
 @EntityListeners({AuditingEntityListener.class})
-public abstract class CreationModificationLog {
+public abstract class CreationLog {
     @CreatedDate
     @Column(updatable = false)
     private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column
-    private LocalDateTime updatedAt;
 }
