+++ aenitto/src/main/java/com/firefighter/aenitto/common/baseEntities/CreationModificationLog.java
@@ -1,7 +1,8 @@
-package com.firefighter.aenitto.domain;
+package com.firefighter.aenitto.common.baseEntities;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.Column;
@@ -12,8 +13,12 @@ import java.time.LocalDateTime;
 @MappedSuperclass
 @Getter
 @EntityListeners({AuditingEntityListener.class})
-public abstract class CreationLog {
+public abstract class CreationModificationLog {
     @CreatedDate
     @Column(updatable = false)
     private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    @Column
+    private LocalDateTime updatedAt;
 }
