--- aenitto/src/main/java/com/firefighter/aenitto/domain/ParticipantRole.java
@@ -1,5 +1,0 @@
-package com.firefighter.aenitto.domain;
-
-
-public enum ParticipantRole {
-}
