+++ aenitto/src/main/java/com/firefighter/aenitto/missions/domain/IndividualMission.java
@@ -1,5 +1,6 @@
-package com.firefighter.aenitto.domain;
+package com.firefighter.aenitto.missions.domain;
 
+import com.firefighter.aenitto.rooms.domain.MemberRoom;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
