+++ aenitto/src/main/java/com/firefighter/aenitto/missions/domain/MissionType.java
@@ -1,4 +1,4 @@
-package com.firefighter.aenitto.domain;
+package com.firefighter.aenitto.missions.domain;
 
 public enum MissionType {
     COMMON, INDIVIDUAL
