+++ aenitto/src/main/java/com/firefighter/aenitto/rooms/domain/ParticipantRole.java
@@ -1,0 +1,5 @@
+package com.firefighter.aenitto.rooms.domain;
+
+
+public enum ParticipantRole {
+}
