+++ aenitto/src/main/java/com/firefighter/aenitto/rooms/domain/RoomState.java
@@ -1,4 +1,4 @@
-package com.firefighter.aenitto.domain;
+package com.firefighter.aenitto.rooms.domain;
 
 public enum RoomState {
     PRE, PROCESSING, POST
